+++ .gitignore
@@ -31,3 +31,5 @@
 *.out
 *.app
 
+server/myapp
+testing/
