+++ bin/dockerSpinup.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-docker run -dit -v ${PWD}:/MG-HTTPS-Server dockerserver bash
+docker run -dit -p 8080:8080 -v ${PWD}:/MG-HTTPS-Server dockerserver bash
