+++ server/Makefile
@@ -1,14 +1,45 @@
-CC 		:= g++
-CFLAGS 	        := -I/usr/local/include -L/usr/local/lib -g
-OBJECTS 	:=
-LIBRARIES 	:= -lhttpserver
+####################### Makefile Template ##########################
+# Compiler settings - Can be customized.
+CC = clang++
+CXXFLAGS = -std=c++17 -Wall
+LDFLAGS = -lhttpserver
 
-.PHONY: all clean
+# Makefile settings - Can be customized.
+APPNAME = myapp
+EXT = .cpp
+SRCDIR = .
+OBJDIR = .
 
-all: hello_world
-
-hello_world: 
-	$(CC) $(CFLAGS) -o hello_world hello_world.cpp $(LIBRARIES)
-
+############## Do not change anything from here downwards! #############
+SRC = $(wildcard $(SRCDIR)/*$(EXT))
+OBJ = $(SRC:$(SRCDIR)/%$(EXT)=$(OBJDIR)/%.o)
+DEP = $(OBJ:$(OBJDIR)/%.o=%.d)
+# UNIX-based OS variables & settings
+RM = rm
+DELOBJ = $(OBJ)
+####################### Targets beginning here #######################
+all: $(APPNAME)
+# Builds the app
+$(APPNAME): $(OBJ)
+	$(CC) $(CXXFLAGS) -o $@ $^ $(LDFLAGS)
+# Creates the dependecy rules
+%.d: $(SRCDIR)/%$(EXT)
+	@$(CPP) $(CFLAGS) $< -MM -MT $(@:%.d=$(OBJDIR)/%.o) >$@
+# Includes all .h files
+-include $(DEP)
+# Building rule for .o files and its .c/.cpp in combination with all .h
+$(OBJDIR)/%.o: $(SRCDIR)/%$(EXT)
+	$(CC) $(CXXFLAGS) -o $@ -c $<
+################### Test ###############################################
+test: all
+	./$(APPNAME)
+################### Cleaning rules for Unix-based OS ###################
+# Cleans complete project
+.PHONY: clean
 clean:
-	rm -f *.o
+	$(RM) $(DELOBJ) $(DEP) $(APPNAME)
+
+# Cleans only all files with the extension .d
+.PHONY: cleandep
+cleandep:
+	$(RM) $(DEP)
